movies/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.13
django-cors-headers==3.5.0
django-rest-auth==0.9.3
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.7
Markdown==3.0.1
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.2
